Moved index page to app directory
